pyproject.toml
[tool.poetry]
name = "course-lm-backend"
version = "0.1.0"
description = ""
authors = ["Bharath Sriraam R R <albharath.1305@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.105.0"
uvicorn = {extras = ["standard"], version = "^0.24.0.post1"}
python-dotenv = "^1.0.0"
langchain = "^0.1.0"
python-jose = "^3.3.0"
black = "^23.12.0"
flake8 = "^6.1.0"
pydantic-settings = "^2.1.0"
supabase = "^2.3.0"
python-multipart = "^0.0.6"
celery = "^5.3.6"
openai = "^1.6.1"
pypdf = "^3.17.3"
tiktoken = "^0.5.2"
pymupdf = "^1.23.8"
langchain-openai = "^0.0.2"
asyncio = "^3.4.3"

[tool.black]
line-length = 130
target-version = ['py38']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | \.eggs
  | \.hg
  | \.mypy_cache
  | \.nox
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

